+++ src/libsystemd/sd-bus/bus-match.c
@@ -22,7 +22,6 @@
 #include "bus-internal.h"
 #include "bus-message.h"
 #include "bus-match.h"
-#include "bus-util.h"
 #include "strv.h"
 
 /* Example:
