+++ src/libsystemd/sd-bus/bus-message.c
@@ -35,7 +35,6 @@
 #include "bus-type.h"
 #include "bus-signature.h"
 #include "bus-gvariant.h"
-#include "bus-util.h"
 
 static int message_append_basic(sd_bus_message *m, char type, const void *p, const void **stored);
