+++ src/libsystemd/sd-bus/bus-objects.c
@@ -26,7 +26,6 @@
 #include "bus-type.h"
 #include "bus-signature.h"
 #include "bus-introspect.h"
-#include "bus-util.h"
 #include "bus-slot.h"
 #include "bus-objects.h"
