+++ src/libsystemd/sd-bus/test-bus-benchmark.c
@@ -28,7 +28,6 @@
 #include "sd-bus.h"
 #include "bus-kernel.h"
 #include "bus-internal.h"
-#include "bus-util.h"
 
 #define MAX_SIZE (2*1024*1024)
