+++ src/libsystemd/sd-bus/test-bus-chat.c
@@ -33,7 +33,6 @@
 #include "bus-error.h"
 #include "bus-match.h"
 #include "bus-internal.h"
-#include "bus-util.h"
 
 static int match_callback(sd_bus_message *m, void *userdata, sd_bus_error *ret_error) {
         log_info("Match triggered! interface=%s member=%s", strna(sd_bus_message_get_interface(m)), strna(sd_bus_message_get_member(m)));
