+++ src/libsystemd/sd-bus/test-bus-cleanup.c
@@ -22,7 +22,6 @@
 #include <stdio.h>
 
 #include "sd-bus.h"
-#include "bus-util.h"
 #include "bus-internal.h"
 #include "bus-message.h"
 #include "refcnt.h"
