+++ src/libsystemd/sd-bus/test-bus-error.c
@@ -20,7 +20,6 @@
 ***/
 
 #include "sd-bus.h"
-#include "bus-error.h"
 #include "bus-util.h"
 #include "errno-list.h"
 #include "bus-common-errors.h"
