+++ src/libsystemd/sd-bus/test-bus-match.c
@@ -24,7 +24,6 @@
 
 #include "bus-match.h"
 #include "bus-message.h"
-#include "bus-util.h"
 #include "bus-slot.h"
 
 static bool mask[32];
