+++ src/libsystemd/sd-bus/test-bus-server.c
@@ -28,7 +28,6 @@
 
 #include "sd-bus.h"
 #include "bus-internal.h"
-#include "bus-util.h"
 
 struct context {
         int fds[2];
